+++ .gitignore
@@ -11,6 +11,7 @@
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
-./octoprint-exporter/octoprint-exporter
 *.bin
 *.retry
+
+configs/config.yaml
